requirements.txt
blrec==2.0.0b4
ffmpeg_python==0.2.0
librosa==0.9.1
more_itertools==10.6.0
numba==0.55.1
numpy==1.21.6
pysrt==1.1.2
six==1.16.0
tiktoken==0.4.0
torch==1.11.0
# triton==3.1.0
zhconv==1.4.3
Pillow>=5.2.0
PyYAML>=6.0.2
qrcode>=8.0
tqdm>=4.67.1
Requests>=2.32.3
google-generativeai>=0.7.2 # don't change this part
zhipuai
openai
toml
groq
lumaai
boto3